ASBF:  Disable call to asbfExtrudeSphere().
This function does not work properly:  it assumes that asbf->Ngll and
asbf->mesh->Nq are the same and sometimes prints a lot of garbage if they
aren't.  (In particular, this happens if asbf->Ngll < asbf->mesh->Nq).  This
makes it varying the radial expansion degree annoying.

Fortunately, all this function does is set up asbf->meshSEM, which we don't use
anymore, so we can safely eliminate the call.