ASBF:  Add second Neumann test case.
This is very similar to the previous test case:  the key difference is that s =
phi^3(phi - pi)^3 instead of phi^2*(phi - pi)^2:  we increase the order of the
zeroes at the poles of the sphere by 1.  This causes q and Lap(q) to be zero
when phi = 0 or phi = pi, killing off the singularities entirely.  Now we can
get down to an error level of 1e-8, as we'd hope.

We will leave the previous test case in (but commented out) for the time being.