ASBF:  Implement separate-then-discretize (Dirichlet case).
We use Chebfun to compute the true eigenfunctions (or, rather, highly accurate
polynomial approximations thereto) offline.  Then, it's just a matter of making
sure we have the right number of quadrature nodes and loading all of the cached
basis information into the right matrices.

R and lambda are hard-coded in the MATLAB file at the moment, as I'm not yet
sure how changes to these parameters affect the eigenvalues (and hence the
eigenfunctions).