ASBF:  Implement piecewise discrete eigenfunction basis (Dirichlet-Dirichlet).
TCEW says we need this because other codes will require the radial solution
expressed in a piecewise polynomial basis.  This code just gets the
eigenfunctions up and working.  To revert to the piecewise polynomial basis
from the eigenfunction basis, we'll need to use the eigenvectors of the
generalized eigenvalue problem; these will eventually need to be stored.

So far, this code is for Dirichlet-Dirichlet BCs only.