ASBF:  Change asbf->BCType to asbf->innerBC and asbf->outerBC.
I misunderstood the role of BCType in the other solvers when I put it in ASBF,
where I had it play a different role; this commit eliminates it from ASBF to
reduce confusion.

Also, get rid of some of the stuff that got chucked in yesterday when debugging
the SEM solver.