shell:  Use scaled cosine for Neumann-Neumann test case.
The previous test case was exactly a degree-3 polynomial in radius, which is a
little too "special".

The new test case is resolved to roughly 1e-8 accuracy with N = 11.  (Maybe
this is a little too high?)  Note that due to even/odd symmetry, one will see
improvement in radial resolution only with every *other* increase in N, e.g.,
the error will improve when going from N = 5 to N = 7 but not from N = 5 to N =
6.