shell:  For ASBF, build preconditioners for each eigenvalue.
This is the right thing to do, and it makes a HUGE difference for
ASBF-MULTIGRID.  Interestingly, it makes almost zero difference for
ASBF-JACOBI (with CONTINUOUS discretization---haven't checked IPDG).

This commit also makes dynamic switching from other preconditioners to
JACOBI an option in the setup file.  Testing (again, with CONTINUOUS
discretization) that this is not necessary for MULTIGRID, even at the
highest-order modes.

One point of concern here is memory usage, especially for MULTIGRID.
This code generates a completely new MULTIGRID preconditioner object for
each eigenvalue, which is wasteful.  We should be able to reuse almost
all of the information for the p-multigrid levels across eigenvalues.
The catch is that this is not true for the AMG levels (as the
explicitly-constructed matrices that we feed to AMG depend on LAMBDA).
It is not immediately clear to me what that "right" way is to reconcile
this with the code structured as it is at the moment.