stokes:  Allocate Stokes vector data as one big long array.
This is what is done in the INS solver.  It makes writing codes that are
independent of the dimension (2D vs. 3D) easier and is probably better
for putting stuff onto the GPU, which we will begin to do in the near
future.

While we're at it, move the functions that operate on the vectors out of
stokesSolve() into their own module.  Again, these functions are
placeholders that will soon be replaced by OCCA kernels.