stokes:  Remove dead host code.
I wrote this as a guide to help me get through writing the corresponding
device code.  It has served its purpose and is now out of date, as it
would require further modification for handling Dirichlet conditions.
Therefore, it should be removed.  If we need to retrieve it, it's always
in Git.  Running on the host can still be accomplished by setting the
THREADING MODEL variable to "Serial" in the setup file.