stokes:  Add options to switch the velocity and pressure preconditioner blocks.
When lambda is large, it can be better to precondition the velocity
block using Jacobi and the pressure block with multigrid.  Some very
casual experimentation thus far suggests that iteration counts are well
controlled, but more testing is needed.

In particular, we need to figure out where the cross-over point is and
what to do in the "in-between" region.