ogs:  Fix apparent indexing error in gatherMany kernels.
This seems to fix the problems we've been having with
ogsGatherScatterMany() on multiple MPI ranks.

But note that it only *seems* to fix the problem:  I do not understand
why this error did not cause issues with just a single MPI rank in play.
It is possible that there is something deeper here that I've missed...